cambiar reglas de validcion y ordenar en carpetas
